+++ summer-manager/summer-manager-service/src/main/java/com/summer/service/ItemService.java
@@ -7,7 +7,7 @@ import com.summer.pojo.TbItem;
  * Created by toy on 6/4/16.
  */
 public interface ItemService {
-    TbItem getItemById(int itemid);
+    TbItem getItemById(long itemid);
 
 
     EasyUIDataGridResult getItemList(int page, int rows);
